+++ src/parser.h
@@ -2,8 +2,7 @@
 #define SRC_PARSER_H_
 
 #include "common.h"  /* mc_frame_t and friends */
-#include <sys/types.h>  /* ssize_t */
 
-int mc_parser_execute(uint8_t* data, ssize_t len, mc_frame_t* frame);
+int mc_parser_execute(uint8_t* data, int len, mc_frame_t* frame);
 
 #endif  /* SRC_PARSER_H_ */
